+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/repository/GudangBarangDb.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface GudangBarangDb extends JpaRepository<GudangBarang, BigInteger> {
+public interface GudangBarangDb extends JpaRepository<GudangBarang, Long> {
 }
