+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/repository/GudangDb.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface GudangDb extends JpaRepository<Gudang, BigInteger> {
+public interface GudangDb extends JpaRepository<Gudang, Long> {
 }
