+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/repository/KaryawanDb.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface KaryawanDb extends JpaRepository<Karyawan, BigInteger> {
+public interface KaryawanDb extends JpaRepository<Karyawan, Long> {
 }
