+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/repository/PermintaanPengirimanBarangDb.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface PermintaanPengirimanBarangDb extends JpaRepository<PermintaanPengirimanBarang, BigInteger> {
+public interface PermintaanPengirimanBarangDb extends JpaRepository<PermintaanPengirimanBarang, Long> {
 }
