+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/service/GudangBarangService.java
@@ -10,5 +10,5 @@ public interface GudangBarangService {
 
     List<GudangBarang> getAllGudangBarang();
 
-    GudangBarang getGudangBarangById(BigInteger id);
+    GudangBarang getGudangBarangById(Long id);
 }
