+++ silogistik2106751070/src/main/java/apap/ti/silogistik2106751070/service/GudangService.java
@@ -11,7 +11,7 @@ public interface GudangService {
 
     List<Gudang> getAllGudang();
 
-    Gudang getGudangById(BigInteger id);
+    Gudang getGudangById(Long id);
 
     long getCount();
